Fix: correct Zombatar avatar part stacking and visibility
Assign each selected part track the draw order from the part layout
table as its render group, so parts stack in a fixed order regardless
of track definition order, and show the "_line" detail track right
above the base track independently, since some parts only exist as a
"_line" track without a base track.

Draw every render group of an attached reanimation so that grouped
part tracks are actually painted.

Share the part layout tables between the widget and the runtime head
setup so both read the same data.

Close #429
